Inputs and Outputs moved from app.py into their own files: inputs.py and outputs.py.
