+++ pywps/app.py
@@ -16,17 +16,11 @@ from lxml import etree
 from pywps.exceptions import InvalidParameterValue, \
     MissingParameterValue, NoApplicableCode, \
     OperationNotSupported, VersionNegotiationFailed, FileSizeExceeded, StorageNotSupported
-from pywps.storage import FileStorage
-from pywps._compat import text_type, StringIO
-from pywps import inout
-from pywps import config
+from pywps._compat import text_type, StringIO, PY2
+from pywps import config, WPS, OWS, NAMESPACES
 from pywps.formats import Format
-
-
-from pywps._compat import PY2
-
-xmlschema_2 = "http://www.w3.org/TR/xmlschema-2/#"
-
+from pywps.inputs import LiteralInput, ComplexInput
+from pywps.outputs import LiteralOutput, ComplexOutput
 
 def xpath_ns(el, path):
     return el.xpath(path, namespaces=NAMESPACES)
@@ -555,480 +549,10 @@ class WPSResponse(object):
         return xml_response(self._construct_doc())
 
 
-class LiteralInput(inout.LiteralInput):
-    """
-    :param identifier: The name of this input.
-    :param data_type: Type of literal input (e.g. `string`, `float`...).
-    """
 
-    def __init__(self, identifier, title, data_type='string', abstract='', metadata=[], uom=[], default='',
-                 min_occurs='1', max_occurs='1', as_reference=False):
-        inout.LiteralInput.__init__(self, identifier=identifier, title=title, abstract=abstract, data_type=data_type)
-        self.metadata = metadata
-        self.uom = uom
-        self.default = default
-        self.min_occurs = min_occurs
-        self.max_occurs = max_occurs
-        self.as_reference = as_reference
 
-    def describe_xml(self):
-        doc = E.Input(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
 
-        doc.attrib['minOccurs'] = self.min_occurs
-        doc.attrib['maxOccurs'] = self.max_occurs
 
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-
-        if self.metadata:
-            doc.append(OWS.Metadata(*self.metadata))
-
-        literal_data_doc = E.LiteralData()
-
-        if self.data_type:
-            literal_data_doc.append(OWS.DataType(self.data_type, reference=xmlschema_2 + self.data_type))
-
-        if self.uom:
-            default_uom_element = self.uom[0].describe_xml()
-            supported_uom_elements = [u.describe_xml() for u in self.uom]
-
-            literal_data_doc.append(
-                E.UOMs(
-                    E.Default(default_uom_element),
-                    E.Supported(*supported_uom_elements)
-                )
-            )
-
-        doc.append(literal_data_doc)
-
-        # TODO: refer to table 29 and 30
-        doc.append(OWS.AnyValue())
-
-        if self.default:
-            doc.append(E.DefaultValue(self.default))
-
-        return doc
-
-    def execute_xml(self):
-        """Render Execute response XML node
-
-        :return: node
-        :rtype: ElementMaker
-        """
-        node = None
-        if self.as_reference:
-            node = self._execute_xml_reference()
-        else:
-            node = self._execute_xml_data()
-
-        doc = WPS.Input(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-        doc.append(node)
-
-        return doc
-
-    def _execute_xml_reference(self):
-        """Return Reference node
-        """
-        doc = WPS.Reference()
-        doc.attrib['{http://www.w3.org/1999/xlink}href'] = self.stream
-        if self.method.upper() == 'POST' or self.method.upper() == 'GET':
-            doc.attrib['method'] = self.method.upper()
-        return doc
-
-    def _execute_xml_data(self):
-        """Return Data node
-        """
-        doc = WPS.Data()
-        literal_doc = WPS.LiteralData(self.data)
-
-        if self.data_type:
-            literal_doc.attrib['dataType'] = self.data_type
-        if self.uom:
-            literal_doc.attrib['uom'] = self.uom
-        doc.append(literal_doc)
-        return doc
-
-
-class ComplexInput(inout.ComplexInput):
-    """
-    :param identifier: The name of this input.
-    :param allowed_formats: Allowed formats for this input. Should be a list of
-                    one or more :class:`~Format` objects.
-    :param data_format: Format of the passed input. Should be :class:`~Format` object
-    """
-
-    def __init__(self, identifier, title, allowed_formats=None, data_format=None, abstract='', metadata=[], min_occurs='1',
-                 max_occurs='1', as_reference=False):
-        inout.ComplexInput.__init__(self, identifier=identifier, title=title, abstract=abstract, data_format=data_format)
-        self.allowed_formats = allowed_formats
-        self.metadata = metadata
-        self.min_occurs = min_occurs
-        self.max_occurs = max_occurs
-        self.as_reference = as_reference
-        self.url = ''
-        self.method = ''
-        self.max_megabytes = int(0)
-
-    def calculate_max_input_size(self):
-        """Calculates maximal size for input file based on configuration
-        and units
-
-        :return: maximum file size bytes
-        """
-        maxSize = config.get_config_value('server', 'maxfilesize')
-        maxSize = maxSize.lower()
-
-        import re
-
-        units = re.compile("[gmkb].*")
-        size = float(re.sub(units, '', maxSize))
-
-        if maxSize.find("g") > -1:
-            size *= 1024
-        elif maxSize.find("m") > -1:
-            size *= 1
-        elif maxSize.find("k") > -1:
-            size /= 1024
-        else:
-            size *= 1
-        self.max_megabytes = size
-
-    def describe_xml(self):
-        default_format_el = self.allowed_formats[0].describe_xml()
-        supported_format_elements = [f.describe_xml() for f in self.allowed_formats]
-        
-        doc = E.Input(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-
-        doc.attrib['minOccurs'] = self.min_occurs
-        doc.attrib['maxOccurs'] = self.max_occurs
-
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-
-        if self.metadata:
-            doc.append(OWS.Metadata(*self.metadata))
-
-        doc.append(
-            E.ComplexData(
-                E.Default(default_format_el),
-                E.Supported(*supported_format_elements)
-            )
-        )
-
-        self.calculate_max_input_size()
-        if self.max_megabytes:
-            doc.attrib['maximumMegabytes'] = str(self.max_megabytes)
-
-        return doc
-    
-    def execute_xml(self):
-        """Render Execute response XML node
-
-
-        :return: node
-        :rtype: ElementMaker
-        """
-        node = None
-        if self.as_reference:
-            node = self._execute_xml_reference()
-        else:
-            node = self._execute_xml_data()
-
-        doc = WPS.Input(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-        doc.append(node)
-
-        return doc
-
-    def _execute_xml_reference(self):
-        """Return Reference node
-        """
-        doc = WPS.Reference()
-        doc.attrib['{http://www.w3.org/1999/xlink}href'] = self.url
-        if self.data_format:
-            if self.data_format.mime_type:
-                doc.attrib['mimeType'] = self.data_format.mime_type
-            if self.data_format.encoding:
-                doc.attrib['encoding'] = self.data_format.encoding
-            if self.data_format.schema:
-                doc.attrib['schema'] = self.data_format.schema
-        if self.method.upper() == 'POST' or self.method.upper() == 'GET':
-            doc.attrib['method'] = self.method.upper()
-        return doc
-
-    def _execute_xml_data(self):
-        """Return Data node
-        """
-        doc = WPS.Data()
-        complex_doc = WPS.ComplexData(self.data)
-
-        if self.data_format:
-            if self.data_format.mime_type:
-                complex_doc.attrib['mimeType'] = self.data_format.mime_type
-            if self.data_format.encoding:
-                complex_doc.attrib['encoding'] = self.data_format.encoding
-            if self.data_format.schema:
-                complex_doc.attrib['schema'] = self.data_format.schema
-        doc.append(complex_doc)
-        return doc
-
-
-class LiteralOutput(inout.LiteralOutput):
-    """
-    :param identifier: The name of this output.
-    :param data_type: Type of literal input (e.g. `string`, `float`...).
-    :param value: Resulting value
-            Should be :class:`~String` object.
-    """
-
-    def __init__(self, identifier, title, data_type='string', abstract='', metadata=[], uom=[]):
-        inout.LiteralOutput.__init__(self, identifier, title=title, data_type=data_type)
-        self.abstract = abstract
-        self.metadata = metadata
-        self.uom = uom
-
-    def describe_xml(self):
-        doc = E.Output(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-
-        for m in self.metadata:
-            doc.append(OWS.Metadata(m))
-
-        literal_data_doc = E.LiteralData()
-
-        if self.data_type:
-            literal_data_doc.append(OWS.DataType(self.data_type, reference=xmlschema_2 + self.data_type))
-
-        if self.uom:
-            default_uom_element = self.uom[0].describe_xml()
-            supported_uom_elements = [u.describe_xml() for u in self.uom]
-
-            literal_data_doc.append(
-                E.UOMs(
-                    E.Default(default_uom_element),
-                    E.Supported(*supported_uom_elements)
-                )
-            )
-
-        return doc
-
-    def execute_xml(self):
-        doc = WPS.Output(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-
-        data_doc = WPS.Data()
-
-        literal_data_doc = WPS.LiteralData(text_type(self.data))
-        literal_data_doc.attrib['dataType'] = self.data_type
-        literal_data_doc.attrib['reference'] = xmlschema_2 + self.data_type
-        if self.uom:
-            default_uom_element = self.uom[0].describe_xml()
-            supported_uom_elements = [u.describe_xml() for u in self.uom]
-
-            literal_data_doc.append(
-                E.UOMs(
-                    E.Default(default_uom_element),
-                    E.Supported(*supported_uom_elements)
-                )
-            )
-        data_doc.append(literal_data_doc)
-
-        doc.append(data_doc)
-
-        return doc
-
-
-class ComplexOutput(inout.ComplexOutput):
-    """
-    :param identifier: The name of this output.
-    :param formats: Possible output formats for this output.
-            Should be list of :class:`~Format` object.
-    :param output_format: Required format for this output.
-            Should be :class:`~Format` object.
-    :param encoding: The encoding of this input or requested for this output
-            (e.g., UTF-8).
-    """
-
-    def __init__(self, identifier, title, formats=None, output_format=None, encoding="UTF-8",
-                 schema=None, abstract='', metadata=[]):
-        inout.ComplexOutput.__init__(self, identifier, title=title, abstract=abstract)
-        self.formats = formats
-        self.metadata = metadata
-
-        self._schema = None
-        self._output_format = None
-        self._encoding = None
-
-        self.as_reference = False
-        self.output_format = output_format
-        self.encoding = encoding
-        self.schema = schema
-        self.storage = None
-
-    @property
-    def output_format(self):
-        """Get output format
-        :rtype: String
-        """
-
-        if self._output_format:
-            return self._output_format
-        else:
-            return ''
-
-    @output_format.setter
-    def output_format(self, output_format):
-        """Set output format
-        """
-        self._output_format = output_format
-
-    @property
-    def encoding(self):
-        """Get output encoding
-        :rtype: String
-        """
-
-        if self._encoding:
-            return self._encoding
-        else:
-            return ''
-
-    @encoding.setter
-    def encoding(self, encoding):
-        """Set output encoding
-        """
-        self._encoding = encoding
-
-    @property
-    def schema(self):
-        """Get output schema
-        :rtype: String
-        """
-
-        return self._schema
-
-    @schema.setter
-    def schema(self, schema):
-        """Set output schema
-        """
-        self._schema = schema
-
-    def describe_xml(self):
-        default_format_el = self.formats[0].describe_xml()
-        supported_format_elements = [f.describe_xml() for f in self.formats]
-
-        doc = E.Output(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-
-        for m in self.metadata:
-            doc.append(OWS.Metadata(*self.metadata))
-
-        doc.append(
-            E.ComplexOutput(
-                E.Default(default_format_el),
-                E.Supported(*supported_format_elements)
-            )
-        )
-
-        return doc
-
-    def execute_xml(self):
-        """Render Execute response XML node
-
-        :return: node
-        :rtype: ElementMaker
-        """
-
-        node = None
-        if self.as_reference:
-            node = self._execute_xml_reference()
-        else:
-            node = self._execute_xml_data()
-
-        doc = WPS.Output(
-            OWS.Identifier(self.identifier),
-            OWS.Title(self.title)
-        )
-        if self.abstract:
-            doc.append(OWS.Abstract(self.abstract))
-        doc.append(node)
-
-        return doc
-
-    def _execute_xml_reference(self):
-        """Return Reference node
-        """
-        doc = WPS.Reference()
-
-        # get_url will create the file and return the url for it
-        self.storage = FileStorage(config)
-        doc.attrib['{http://www.w3.org/1999/xlink}href'] = self.get_url()
-
-        if self.data_format:
-            if self.data_format.mime_type:
-                doc.attrib['mimeType'] = self.data_format.mime_type
-            if self.data_format.encoding:
-                doc.attrib['encoding'] = self.data_format.encoding
-            if self.data_format.schema:
-                doc.attrib['schema'] = self.data_format.schema
-        return doc
-
-    def _execute_xml_data(self):
-        """Return Data node
-        """
-        doc = WPS.Data()
-
-        if self.data is None:
-            complex_doc = WPS.ComplexData()
-        else:
-            complex_doc = WPS.ComplexData(self.data)
-
-        if self.data_format:
-            if self.data_format.mime_type:
-                complex_doc.attrib['mimeType'] = self.data_format.mime_type
-            if self.data_format.encoding:
-                complex_doc.attrib['encoding'] = self.data_format.encoding
-            if self.data_format.schema:
-                complex_doc.attrib['schema'] = self.data_format.schema
-        doc.append(complex_doc)
-        return doc
-
-
-class BoundingBoxOutput(object):
-    """bounding box output
-    """
-    # TODO: BoundingBoxOutput
-    pass
 
 
 class Process(object):
